scripts/requirements-mine.txt
whisperx>=3.1.1
openai-whisper>=20231117
resemblyzer>=0.1.3
soundfile>=0.12.1
numpy>=1.24
torch>=2.1
torchaudio>=2.1
pyannote.audio>=3.1
yt-dlp>=2024.1.1
